api change

+++ admin-api-lib/src/admin_api_lib/apis/admin_api.py
@@ -2,14 +2,36 @@
 
 # coding: utf-8
 
+from typing import Dict, List  # noqa: F401
 import importlib
 import pkgutil
 
+from admin_api_lib.apis.admin_api_base import BaseAdminApi
 from fastapi import APIRouter, Path, Request, Response, UploadFile  # noqa: F401
 
 import admin_api_lib.impl
-from admin_api_lib.apis.admin_api_base import BaseAdminApi
+
+from fastapi import (  # noqa: F401
+    APIRouter,
+    Body,
+    Cookie,
+    Depends,
+    Form,
+    Header,
+    HTTPException,
+    Path,
+    Query,
+    Response,
+    Security,
+    status,
+)
+
+from admin_api_lib.models.extra_models import TokenModel  # noqa: F401
+from pydantic import Field, StrictBytes, StrictStr
+from typing import Any, List, Tuple, Union
+from typing_extensions import Annotated
 from admin_api_lib.models.document_status import DocumentStatus
+from admin_api_lib.models.upload_source import UploadSource
 
 
 router = APIRouter()
@@ -43,6 +65,8 @@ async def delete_document(
     -------
     None
     """
+    if not BaseAdminApi.subclasses:
+        raise HTTPException(status_code=500, detail="Not implemented")
     return await BaseAdminApi.subclasses[0]().delete_document(identification)
 
 
@@ -73,6 +97,8 @@ async def document_reference_id_get(
     Response
         The response object containing the document reference details.
     """
+    if not BaseAdminApi.subclasses:
+        raise HTTPException(status_code=500, detail="Not implemented")
     return await BaseAdminApi.subclasses[0]().document_reference_id_get(identification)
 
 
@@ -94,39 +120,13 @@ async def get_all_documents_status() -> list[DocumentStatus]:
     list[DocumentStatus]
         A list containing the status of all documents.
     """
+    if not BaseAdminApi.subclasses:
+        raise HTTPException(status_code=500, detail="Not implemented")
     return await BaseAdminApi.subclasses[0]().get_all_documents_status()
 
 
 @router.post(
-    "/load_confluence",
-    responses={
-        200: {"description": "Loading from confluence is successful"},
-        423: {
-            "description": (
-                "if the confluence loader is already processing a request,"
-                "no further requests are possible. The user needs to wait,"
-                "till the preliminary request finished processing."
-            )
-        },
-        500: {"description": "Internal Server Error"},
-        501: {"description": "The confluence loader is not set up"},
-    },
-    tags=["admin"],
-    response_model_by_alias=True,
-)
-async def load_confluence_post() -> None:
-    """
-    Asynchronously loads a Confluence space.
-
-    Returns
-    -------
-    None
-    """
-    return await BaseAdminApi.subclasses[0]().load_confluence_post()
-
-
-@router.post(
-    "/upload_documents",
+    "/upload_source",
     responses={
         200: {"description": "ok"},
         400: {"description": "Bad request"},
@@ -136,22 +136,10 @@ async def load_confluence_post() -> None:
     tags=["admin"],
     response_model_by_alias=True,
 )
-async def upload_documents_post(
-    body: UploadFile,
-    request: Request,
+async def upload_source(
+    upload_source: Annotated[UploadSource, Field(description="The source to upload.")] = Body(None, description="The source to upload."),
 ) -> None:
-    """
-    Asynchronously uploads user-selected source documents.
-
-    Parameters
-    ----------
-    body : UploadFile
-        The file object containing the source documents to be uploaded.
-    request : Request
-        The request object containing metadata about the upload request.
-
-    Returns
-    -------
-    None
-    """
-    return await BaseAdminApi.subclasses[0]().upload_documents_post(body, request)
+    """Uploads user selected source."""
+    if not BaseAdminApi.subclasses:
+        raise HTTPException(status_code=500, detail="Not implemented")
+    return await BaseAdminApi.subclasses[0]().upload_source(upload_source)
